refactor: extract width custom hooks
